Improved panel view structure
