+++ assets/dev/js/editor/layouts/panel/layout.js
@@ -1,0 +1,223 @@
+var EditModeItemView = require( 'elementor-layouts/edit-mode' ),
+	PanelLayoutView;
+
+PanelLayoutView = Marionette.LayoutView.extend( {
+	template: '#tmpl-elementor-panel',
+
+	id: 'elementor-panel-inner',
+
+	regions: {
+		content: '#elementor-panel-content-wrapper',
+		header: '#elementor-panel-header-wrapper',
+		footer: '#elementor-panel-footer',
+		modeSwitcher: '#elementor-mode-switcher'
+	},
+
+	pages: {},
+
+	childEvents: {
+		'click:add': function() {
+			this.setPage( 'elements' );
+		},
+		'editor:destroy': function() {
+			this.setPage( 'elements' );
+		}
+	},
+
+	currentPageName: null,
+
+	currentPageView: null,
+
+	_isScrollbarInitialized: false,
+
+	initialize: function() {
+		this.initPages();
+	},
+
+	buildPages: function() {
+		var pages = {
+			elements: {
+				view: require( 'elementor-panel/pages/elements/elements' ),
+				title: '<img src="' + elementor.config.assets_url + 'images/logo-panel.svg">'
+			},
+			editor: {
+				view: require( 'elementor-panel/pages/editor' )
+			},
+			menu: {
+				view: elementor.modules.layouts.panel.pages.menu.Menu,
+				title: '<img src="' + elementor.config.assets_url + 'images/logo-panel.svg">'
+			},
+			colorScheme: {
+				view: require( 'elementor-panel/pages/schemes/colors' )
+			},
+			typographyScheme: {
+				view: require( 'elementor-panel/pages/schemes/typography' )
+			},
+			colorPickerScheme: {
+				view: require( 'elementor-panel/pages/schemes/color-picker' )
+			}
+		};
+
+		var schemesTypes = Object.keys( elementor.schemes.getSchemes() ),
+			disabledSchemes = _.difference( schemesTypes, elementor.schemes.getEnabledSchemesTypes() );
+
+		_.each( disabledSchemes, function( schemeType ) {
+			var scheme  = elementor.schemes.getScheme( schemeType );
+
+			pages[ schemeType + 'Scheme' ].view = require( 'elementor-panel/pages/schemes/disabled' ).extend( {
+				disabledTitle: scheme.disabled_title
+			} );
+		} );
+
+		return pages;
+	},
+
+	initPages: function() {
+		var pages;
+
+		this.getPages = function( page ) {
+			if ( ! pages ) {
+				pages = this.buildPages();
+			}
+
+			return page ? pages[ page ] : pages;
+		};
+
+		this.addPage = function( pageName, pageData ) {
+			if ( ! pages ) {
+				pages = this.buildPages();
+			}
+
+			pages[ pageName ] = pageData;
+		};
+	},
+
+	getHeaderView: function() {
+		return this.getChildView( 'header' );
+	},
+
+	getFooterView: function() {
+		return this.getChildView( 'footer' );
+	},
+
+	getCurrentPageName: function() {
+		return this.currentPageName;
+	},
+
+	getCurrentPageView: function() {
+		return this.currentPageView;
+	},
+
+	setPage: function( page, title, viewOptions ) {
+		if ( 'elements' === page && ! elementor.userCan( 'design' ) ) {
+			var pages = this.getPages();
+			if ( pages.hasOwnProperty( 'page_settings' ) ) {
+				page = 'page_settings';
+			}
+		}
+		var pageData = this.getPages( page );
+
+		if ( ! pageData ) {
+			throw new ReferenceError( 'Elementor panel doesn\'t have page named \'' + page + '\'' );
+		}
+
+		if ( pageData.options ) {
+			viewOptions = _.extend( pageData.options, viewOptions );
+		}
+
+		var View = pageData.view;
+
+		if ( pageData.getView ) {
+			View = pageData.getView();
+		}
+
+		this.currentPageName = page;
+
+		this.currentPageView = new View( viewOptions );
+
+		this.showChildView( 'content', this.currentPageView );
+
+		this.getHeaderView().setTitle( title || pageData.title );
+
+		this
+			.trigger( 'set:page', this.currentPageView )
+			.trigger( 'set:page:' + page, this.currentPageView );
+	},
+
+	openEditor: function( model, view ) {
+		var currentPageName = this.getCurrentPageName();
+
+		if ( 'editor' === currentPageName ) {
+			var currentPageView = this.getCurrentPageView(),
+				currentEditableModel = currentPageView.model;
+
+			if ( currentEditableModel === model ) {
+				return;
+			}
+		}
+
+		var elementData = elementor.getElementData( model );
+
+		this.setPage( 'editor', elementor.translate( 'edit_element', [ elementData.title ] ), {
+			model: model,
+			controls: elementor.getElementControls( model ),
+			editedElementView: view
+		} );
+
+		var action = 'panel/open_editor/' + model.get( 'elType' );
+
+		// Example: panel/open_editor/widget
+		elementor.hooks.doAction( action, this, model, view );
+
+		// Example: panel/open_editor/widget/heading
+		elementor.hooks.doAction( action + '/' + model.get( 'widgetType' ), this, model, view );
+	},
+
+	onBeforeShow: function() {
+		var PanelFooterItemView = require( 'elementor-layouts/panel/footer' ),
+			PanelHeaderItemView = require( 'elementor-layouts/panel/header' );
+
+		// Edit Mode
+		this.showChildView( 'modeSwitcher', new EditModeItemView() );
+
+		// Header
+		this.showChildView( 'header', new PanelHeaderItemView() );
+
+		// Footer
+		this.showChildView( 'footer', new PanelFooterItemView() );
+
+		// Added Editor events
+		this.updateScrollbar = _.throttle( this.updateScrollbar, 100 );
+
+		this.getRegion( 'content' )
+			.on( 'before:show', this.onEditorBeforeShow.bind( this ) )
+			.on( 'empty', this.onEditorEmpty.bind( this ) )
+			.on( 'show', this.updateScrollbar.bind( this ) );
+
+		// Set default page to elements
+		this.setPage( 'elements' );
+	},
+
+	onEditorBeforeShow: function() {
+		_.defer( this.updateScrollbar.bind( this ) );
+	},
+
+	onEditorEmpty: function() {
+		this.updateScrollbar();
+	},
+
+	updateScrollbar: function() {
+		var $panel = this.content.$el;
+
+		if ( ! this._isScrollbarInitialized ) {
+			$panel.perfectScrollbar();
+			this._isScrollbarInitialized = true;
+
+			return;
+		}
+
+		$panel.perfectScrollbar( 'update' );
+	}
+} );
+
+module.exports = PanelLayoutView;
